backend-python/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
backports.tarfile==1.2.0
bidict==0.23.1
blinker==1.7.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cleo==2.1.0
click==8.1.8
colorama==0.4.6
coverage==7.8.0
crashtest==0.4.1
cryptography==41.0.7
dbus-python==1.3.2
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
distro-info==1.7+build1
docker==7.1.0
dulwich==0.22.8
exceptiongroup==1.2.2
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
FLAML==2.3.5
frozenlist==1.6.0
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
jiter==0.10.0
keyring==25.6.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.4.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.83.0
packaging==25.0
pbs-installer==2025.4.9
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.5.0
poetry==2.1.0
poetry-core==2.1.0
propcache==0.3.1
pyautogen==0.2.35
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
pypi-timemachine==0.2
pyproject-api==1.9.0
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-apt==2.7.7+ubuntu4
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
ssh-import-id==5.11
starlette==0.46.2
termcolor==3.1.0
tiktoken==0.9.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tox==4.25.0
tqdm==4.67.1
trove-classifiers==2025.5.1.12
typing-inspection==0.4.1
typing_extensions==4.13.2
unattended-upgrades==0.1
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.1
wadllib==1.3.6
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
